Add files via upload
This code is unnecessarily complicated and modified based on the other m file. While the other m code generates the dat file for 3D geometry with the height in z axis in Palabos and the fluid should flow with x axis direction, this code generates the dat file that Y axis in Palabos as the domian height, the fluid flows as z axis direction.